+++ priv/repo/migrations/20191017225002_drop_websub_tables.exs
@@ -1,8 +1,10 @@
 defmodule Pleroma.Repo.Migrations.DropWebsubTables do
   use Ecto.Migration
 
-  def change do
+  def up do
     drop_if_exists(table(:websub_client_subscriptions))
     drop_if_exists(table(:websub_server_subscriptions))
   end
+
+  def down, do: :noop
 end