+++ test/web/streamer/streamer_test.exs
@@ -215,7 +215,7 @@ defmodule Pleroma.Web.StreamerTest do
     test "it sends message if recipients invalid and thread containment is enabled but user's thread containment is disabled" do
       Pleroma.Config.put([:instance, :skip_thread_containment], false)
       author = insert(:user)
-      user = insert(:user, info: %{skip_thread_containment: true})
+      user = insert(:user, skip_thread_containment: true)
       User.follow(user, author, "accept")
 
       activity =